requirements.txt
Django~=5.1.1
django-seed~=0.3.1
psycopg2-binary~=2.9.9
djangorestframework~=3.14.0
